refactor: move http client into its own module
