+++ kmamiz_data_processor/src/http_client/zipkin.rs
@@ -1,0 +1,15 @@
+use reqwest::Client;
+
+use crate::env::Env;
+
+#[derive(Debug)]
+pub struct ZipkinClient<'a> {
+    client: Client,
+    zipkin_url: &'a String,
+}
+
+impl<'a> ZipkinClient<'a> {
+    pub fn new(env: &'a Env) -> Self {
+        ZipkinClient { client: Client::new(), zipkin_url: &env.zipkin_url }
+    }
+}
